clear host settings on delete
Clearing the machine reference breaks the link to the host, but does
not cause it to be deprovisioned. Clear the image so it is
deprovisioned. Clear the user data since we set that when we allocated
the host and it should be reset for the next use. Turn the host off by
changing the "online" flag to false so that after deprovisioning the
host is powered off.