fixes bug that chose the same host every time
A pointer bug caused the actuator to always choose the last host in the slice
of hosts, regardless of whether it was available or not.

The unit tests could have caught this bug, except that the test data happened
to be arranged such that the expected correct host to choose was always the
last one in the slice.

fixes #51