Don't remove MachineRef from Host before Machine deleted
Ensure that a Machine associated with a Host remains associated at least
until the Host has been deprovisioned when the Machine is being deleted.
This will allow the Machine controller to continue to receive events
when the Host changes (which will reduce latency when detecting the
completion of deprovisioning) and ensure that another Machine cannot
claim the Host until the current one is ready to release it.

Signed-off-by: Zane Bitter <zbitter@redhat.com>