only set image and user data when we actually want to provision
We only want to update the image setting if the host does not already
have a consumer and image.

A consumer without an image is "externally provisioned" and assigning
the image will trigger reprovisioning. Since the control plane nodes
will be configured this way, reprovisioning will take them offline and
try to turn them into workers.

A host with an existing image is also already provisioned and upgrades
are not supported at this time. To re-provision a host, we must fully
deprovision it and then provision it again.

Fixes #93