Handle deleting Machines with externally provisioned Hosts
If a Host is in an "externally provisioned" state (as e.g. masters
generally are once Host objects are created for them), then we have no
control over the Host's provisioning state. Therefore when deleting a
Machine with an externally provisioned Host, we should wait only until
the Host has been powered off rather than for it to be
deprovisioned (which will never happen).

Closes #95

Signed-off-by: Zane Bitter <zbitter@redhat.com>