kustomization.yaml: Use patchesStrategicMerge.
Change "patches" to "patchesStrategicMerge".  The use of "patches" has
been deprecated since v1.0.9:

https://github.com/kubernetes-sigs/kustomize/blob/v1.0.9/pkg/types/kustomization.go#L129

With kustomize v3, this fails with a message similar to the one seen
in this issue:

https://github.com/kubernetes-sigs/kustomize/issues/1373

This change should not result in any change in behavior, but makes
this kustomization.yaml no longer use a deprecated field.