fixes stuck deletion when host has no Status
If a Machine is being deleted and is associated with a BareMetalHost that has
no status, the deletion logic would fail to identify that the host is
unprovisioned. It would requeue forever, waiting for the host to reach an
unprovisioned state.

This can happen if there are BareMetalHosts without a running
baremetal-operator, and Machines with this provider running.