Update dependencies to solve  googleapis/gnostic openapiv2 issue
client-go and kube-openapi required two incompatible versions of
gnostic. Updating to v0.19.0-alpha.2 solved the conflict

Due to the upgrade, the namespaces need to be created when running
the tests, so a single namespace "myns" is used accross the tests
and created in the test suite setup for the baremetal folder.