Remove 'run' make target
This PR removes run make target from Makefile since after the uplift to
v1alpha5 version, it is not feasible to run go code only anymore. For
local development and testing, Tilt is recommended instead.

This PR also updates the BMH CRDs to keep it up to date.

This PR also uplifts the envsubst module in tools.