Add re-pivoting e2e test
The main goal of this commit to implement the re-pivoting e2e
test.

This commit also renames the e2e test functions to be aligned
with GO naming conventions.

This commit also adds the process of setting the nodeReuse
varialbe of metal3MachineTemplates to "false" in order to
fix target cluster cleanup issue that is present the value is "true"
isntead of "false".