Reconcile template controller on changes to Metal3Machine
Normally, automated cleaning mode (either disabled or
metadata) is replicated from a Metal3MachineTemplate spec
to Metal3Machine spec. It is allowed to patch Metal3Machine's
automtedCleaningMode directly, and if that's the case CAPM3
controller should notice is, request a reconciliation for the
Metal3MachineTemplate and reset the automatedCleaningMode on
the Metal3Machine to the same as in Metal3MachineTemplate.
This ensures that the source of the truth is always in the
template.