Exclude rules G107 for gosec
Currently the way URL is handled in downloadFile() function of test/e2e/commonn.go is causing an error, G107 (CWE-88): Potential HTTP request made with variable url. The only way to solve this is to call http.Get with a const URL value. Since this function is called from multiple files, its not possible to call it using const value of URL. Which is why currently we are excluding this rule. Once we find a more appropriate way, we shold remove this exclusion.