Fix e2e waiting for status
In the e2e tests we often have to wait for BMHs/M3Ms/Machines to be in
some specific state. In some places we had mistakenly not added any
conditions to the Eventually call, or used it in a not supported way.
This commit makes the use of Eventually more consistent and fixes the
missing conditions.