add explicit securitycontexts to controller
It is always good to not rely on the defaults, but be explicit.
Set explicit, secure securityContext for the capm3 controller manager
deployment and containers.

Setting explicit securityContext has its downsides as well, for tilt.
Tilt's live update cannot handle securityContext which sets the user
as non-root, as it requires root to deploy the binaries on rebuild.
For this, we add a simple Python wrapper that strips CAPI's security
contexts away (they'll include them starting 1.4 release), and add
similar strip function to Tiltfile to handle CAPM3 securityContexts.