hack: update ensure-kind.sh
Take ensure-kind.sh from CAPI, that has added support for Mac (Darwin)
and also raised minimum kind version to 0.17.0.

Also, fix the script not to fail if GOPATH_BIN in not in user's path
before running the script, but print a warning and then use full path
for the kind so we can do version check properly.