Bump golangci-lint to v1.52.1
With this commit we also modify the way we install golangci-lint. Previously we were downloading and installing it through gomodules, however newer versions are not woorking properly with that installation method. Now, with this patch we will be installing it through the workflow file and the ensure-golangci-lint.sh which downloads the binary from github release. We also fix the lint errors here.

Co-authored by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>