add hack/verify-release.sh script
Script that can be used to verify release content. If release tag
has not been created, it can be used to verify go.mod modules are
the exact same version and scanning for vulnerabilities.

Add a mention of the script to the releasing.md, and fix the line
length issues there as well.