Parameterize scalability test
Scalability test is currently running with hardcoded number of clusters and number of clusters provisioned in parallel (both are set to 5). Also, the number of clusters cannot be lower than the number of fake nodes set by dev-env, but so far we have not taken it into account.

We should parameterize these numbers, and use the NUM_NODES param from dev-env as number of clusters. This could change in the future, when we include multi-node clusters into scalability test, but that won't likely happen in a near future.

Signed-off-by: Huy Mai <huy.mai@est.tech>