Refactor setting ProviderID
We are responsible for setting ProviderIDs on Nodes.

Currently, there are two scenarios how this can happen:

- The Node will have a ProviderID already and we need to find the Node,
  read the ProviderID and set it on the Metal3Machine.  This scenario is
  is in effect when a cloud provider is enabled.
- The Node will have a label which we can use to find it and then we can
  set a ProviderID on both the Metal3Machine and the Node.  This
  scenario is in effect when a kubeadm or similar is configured.

Here we add a new method for finding the Node: the hostname.  When the
creation of the Node and its approval into the cluster depends on some
other component, we need to set the ProviderID on the Metal3Machine and
set it to ready.  The external process will then approve the Node and
then we can start looking for it.  Once we find it, we will set the
ProviderID on it.  In order to support this workflow, we change the show
stopping check to include "do we have a node?".  Finally, we add a new
method of finding the Node: by hostname.

The refactor is in the reconcileNormal method and it makes the above
scenarios explicit.

Signed-off-by: Honza Pokorny <honza@redhat.com>