fix for 1.10->1.11 upgrade issue
This commit:
 - Implements a fallback Metal3Data retrieval process as part of the
   Metal3DataTemplate controller's Metal3DataClaim and Metal3Data deletion
   process
   -- The new process allows the controller to retrieve the Metal3Data based
      on the Metal3DataClaim's render information that supposed to point
      directly to the Metal3Data
 - Adds related unit test
 - Extends the existing "UpdateDatas" unit test table evaluation process
   to take into consideration that when a claim is deleted the related
   data has to be deleted too

Original issue:

If the templateReference was used before upgrade to 1.11 and Metal3Data was
generated based on the `templateReference name + Metal3DataClaim index`
instead of the `template name + data claim index` then after the upgrade during
the first new KCP or Metal3MachineDeployment rollout CAPM3 was unable to find
the old Metal3Data because after upgrade to 1.11 CAPM3 was only looking
for the Metal3Data based on `template name + data claim index` only while old
Metal3Data  was named based on `templateReference + index`.

In CAPM3 technically Metal3Data is allowed to get orphaned in case
CAPM3 can't find them based on the information in the Metal3DataClaim and
the related Metal3DataTemplate so the Metal3Data with names based on the
deprecated templateRefference got orphaned and then they were holding onto
IPClaims.

The presence of the orphaned Metal3Data in environments where IP
pre-allocation was used resulted in the old Metal3Data holding onto the pre
allocated IP claims thus the during the rollout the newly created Machies got
stuck as they were unable to re-claim the IPs.

Signed-off-by: Adam Rozman <adam.rozman@est.tech>